refactor: 데이터 사용 util 생성
로컬 DB는 util클래스를 만들어 사용하도록 변경
